pyproject.toml
[project]
name = "feedback-report-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Aryan",email = "aryanbhardwaj3166@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "flower (>=2.0.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
